requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
anyio=3.3.0=py37h89c1867_0
argcomplete=1.12.3=pyhd8ed1ab_2
argon2-cffi=20.1.0=py37h5e8e339_2
async_generator=1.10=py_0
attrs=21.2.0=pyhd8ed1ab_0
babel=2.9.1=pyh44b312d_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=py_2
backports.functools_lru_cache=1.6.4=pyhd8ed1ab_0
blas=1.0=mkl
bleach=4.1.0=pyhd8ed1ab_0
bottleneck=1.3.2=py37heb32a55_1
brotlipy=0.7.0=py37h5e8e339_1001
bzip2=1.0.8=h7b6447c_0
ca-certificates=2021.5.30=ha878542_0
certifi=2021.5.30=py37h89c1867_0
cffi=1.14.6=py37hc58025e_0
chardet=4.0.0=py37h89c1867_1
charset-normalizer=2.0.0=pyhd8ed1ab_0
cryptography=3.4.7=py37h5d9358c_0
cudatoolkit=10.2.89=hfd86e86_1
debugpy=1.4.1=py37hcd2ae1e_0
decorator=5.0.9=pyhd8ed1ab_0
defusedxml=0.7.1=pyhd8ed1ab_0
entrypoints=0.3=pyhd8ed1ab_1003
ffmpeg=4.3=hf484d3e_0
freetype=2.10.4=h5ab3b9f_0
geographiclib=1.52=pypi_0
geopy=2.2.0=pypi_0
gmp=6.2.1=h2531618_2
gnutls=3.6.15=he1e5248_0
idna=3.1=pyhd3deb0d_0
importlib-metadata=4.7.1=py37h89c1867_1
importlib_metadata=4.7.1=hd8ed1ab_1
intel-openmp=2021.3.0=h06a4308_3350
ipykernel=6.2.0=py37h6531663_0
ipython=7.26.0=py37h6531663_0
ipython_genutils=0.2.0=py_1
jedi=0.18.0=py37h89c1867_2
jinja2=3.0.1=pyhd8ed1ab_0
joblib=1.0.1=pyhd8ed1ab_0
jpeg=9b=h024ee3a_2
json5=0.9.5=pyh9f0ad1d_0
jsonschema=3.2.0=pyhd8ed1ab_3
jupyter_client=7.0.1=pyhd8ed1ab_0
jupyter_core=4.7.1=py37h89c1867_0
jupyter_server=1.10.2=pyhd8ed1ab_0
jupyterlab=3.1.9=pyhd8ed1ab_0
jupyterlab_pygments=0.1.2=pyh9f0ad1d_0
jupyterlab_server=2.7.2=pyhd8ed1ab_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.35.1=h7274673_9
libblas=3.9.0=11_linux64_mkl
libcblas=3.9.0=11_linux64_mkl
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h5101ec6_17
libgfortran-ng=11.2.0=h69a702a_8
libgfortran5=11.2.0=h5c6108e_8
libgomp=9.3.0=h5101ec6_17
libiconv=1.15=h63c8f33_5
libidn2=2.3.2=h7f8727e_0
liblapack=3.9.0=11_linux64_mkl
libpng=1.6.37=hbc83047_0
libsodium=1.0.18=h36c2ea0_1
libstdcxx-ng=9.3.0=hd4cf53a_17
libtasn1=4.16.0=h27cfd23_0
libtiff=4.2.0=h85742a9_0
libunistring=0.9.10=h27cfd23_0
libuv=1.40.0=h7b6447c_0
libwebp-base=1.2.0=h27cfd23_0
lz4-c=1.9.3=h295c915_1
markupsafe=2.0.1=py37h5e8e339_0
matplotlib-inline=0.1.2=pyhd8ed1ab_2
mistune=0.8.4=py37h5e8e339_1004
mkl=2021.3.0=h06a4308_520
mkl-service=2.4.0=py37h7f8727e_0
mkl_fft=1.3.0=py37h42c9631_2
mkl_random=1.2.2=py37h51133e4_0
nbclassic=0.3.1=pyhd8ed1ab_1
nbclient=0.5.4=pyhd8ed1ab_0
nbconvert=6.1.0=py37h89c1867_0
nbformat=5.1.3=pyhd8ed1ab_0
ncurses=6.2=he6710b0_1
nest-asyncio=1.5.1=pyhd8ed1ab_0
nettle=3.7.3=hbbd107a_1
ninja=1.10.2=hff7bd54_1
notebook=6.4.3=pyha770c72_0
numexpr=2.7.3=py37h22e1b3c_1
numpy=1.20.3=py37hf144106_0
numpy-base=1.20.3=py37h74d4b33_0
olefile=0.46=py37_0
openh264=2.1.0=hd408876_0
openjpeg=2.3.0=h05c96fa_1
openssl=1.1.1l=h7f8727e_0
packaging=21.0=pyhd8ed1ab_0
pandas=1.3.2=py37h8c16a72_0
pandoc=2.14.2=h7f98852_0
pandocfilters=1.4.2=py_1
parso=0.8.2=pyhd8ed1ab_0
pexpect=4.8.0=pyh9f0ad1d_2
pickleshare=0.7.5=py_1003
pillow=8.3.1=py37h2c7a002_0
pip=21.0.1=py37h06a4308_0
prometheus_client=0.11.0=pyhd8ed1ab_0
prompt-toolkit=3.0.20=pyha770c72_0
ptyprocess=0.7.0=pyhd3deb0d_0
pycparser=2.20=pyh9f0ad1d_2
pygments=2.10.0=pyhd8ed1ab_0
pyopenssl=20.0.1=pyhd8ed1ab_0
pyparsing=2.4.7=pyh9f0ad1d_0
pyrsistent=0.17.3=py37h5e8e339_2
pysocks=1.7.1=py37h89c1867_3
python=3.7.11=h12debd9_0
python-dateutil=2.8.2=pyhd8ed1ab_0
python_abi=3.7=2_cp37m
pytorch=1.9.0=py3.7_cuda10.2_cudnn7.6.5_0
pytz=2021.1=pyhd8ed1ab_0
pyzmq=19.0.2=py37hac76be4_2
readline=8.1=h27cfd23_0
requests=2.26.0=pyhd8ed1ab_0
requests-unixsocket=0.2.0=py_0
scikit-learn=0.24.2=py37h18a542f_0
scipy=1.6.3=py37h29e03ee_0
send2trash=1.8.0=pyhd8ed1ab_0
setuptools=52.0.0=py37h06a4308_0
six=1.16.0=pyh6c4a22f_0
sniffio=1.2.0=py37h89c1867_1
sqlite=3.36.0=hc218d9a_0
terminado=0.11.1=py37h89c1867_0
testpath=0.5.0=pyhd8ed1ab_0
threadpoolctl=2.2.0=pyh8a188c0_0
tk=8.6.10=hbc83047_0
torchaudio=0.9.0=py37
torchvision=0.10.0=py37_cu102
tornado=6.1=py37h5e8e339_1
traitlets=5.0.5=py_0
typing_extensions=3.10.0.0=pyha770c72_0
urllib3=1.26.6=pyhd8ed1ab_0
wcwidth=0.2.5=pyh9f0ad1d_2
webencodings=0.5.1=py_1
websocket-client=0.57.0=py37h89c1867_4
wheel=0.37.0=pyhd3eb1b0_0
xz=5.2.5=h7b6447c_0
zeromq=4.3.4=h9c3ff4c_0
zipp=3.5.0=pyhd8ed1ab_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.9=haebb681_0
